+++ src/lib.rs
@@ -29,7 +29,7 @@
 #![cfg_attr(feature="clippy", feature(plugin))]
 #![cfg_attr(feature="clippy", plugin(clippy))]
 #![cfg_attr(feature="clippy", warn(clippy))]
-#![cfg_attr(feature="clippy", allow(if_not_else))]
+#![cfg_attr(feature="clippy", allow(if_not_else, similar_names))]
 
 #[macro_use]
 extern crate lazy_static;